♻️ Create a console collection
